Address review of wasm heap hardening
