+++ src/compat/web-ifc/ifc_api.ts
@@ -7,7 +7,7 @@ import { ProgressCallback } from '../../core/progress'
 import { ModelInfo } from '../../core/progress_log'
 import Environment from '../../utilities/environment'
 import * as glmatrix from 'gl-matrix'
-import { wasmAddress } from '../../core/wasm_heap'
+import { wasmAddress } from '../../core/wasm_address'
 import { StepExternalByteStore } from '../../step/step_buffer_provider'
 import { IfcApiModelPassthrough } from './ifc_api_model_passthrough'
 import { IfcApiModelPassthroughFactory } from './ifc_api_model_passthrough_factory'
